# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tribe test",
        "description": "Tribe test API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
        "/api/top": {
            "get": {
                "summary": "Show top results list.",
                "description": "The members top results list",
                "operationId": "GetTopList",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Member's email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The members top results list"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/results": {
            "post": {
                "summary": "Store a newly created resource in storage.",
                "description": "Create member's result by email",
                "operationId": "Create result",
                "parameters": [
                    {
                        "name": "milliseconds",
                        "in": "query",
                        "description": "Result in milliseconds",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Member's email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The members top results list"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        }
    }
}